[#1] Refactor: 기능 추상화 및 불필요 코드 제거
